made progress on shaders and abstracted code

+++ src/fragment_shader.glsl
@@ -1,6 +1,5 @@
 #version 300 es
-
-precision highp float;
+precision lowp float;
 
 out vec4 finalColor;
